Implement session manager in guided metadata page
